Drug-stock predictor/Desktop/Drug-stock-predictor/med-predict/py-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pyjwt==2.8.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7
python-jose[cryptography]==3.3.0
email-validator==2.1.0
httpx==0.25.2
alembic==1.13.0
gunicorn==21.2.0
